// Desktop/shash/src/services/enum.ts
export enum YURISH_NAVBATI_ENUM {
    OQ = 'OQ',
    QORA = 'QORA',
}

export enum DONA_TURI_ENUM {
    OQ = 'OQ',
    QORA = 'QORA',
    BOSH_MAYDON = 'BOSH_MAYDON',
}

export enum DONA_ORNI_ENUM {
    ODDIY = 'ODDIY',
    QIROL = 'QIROL',
}

export enum MAYDON_TURI_ENUM {
    ODATIY_MAYDON = 'ODATIY_MAYDON',
    OLINADIGAN_MAYDON = 'OLINADIGAN_MAYDON',
    YURILADIGAN_MAYDON = 'YURILADIGAN_MAYDON',
    YURILAYOTGAN_MAYDON = 'YURILAYOTGAN_MAYDON',
}